requirements.txt
Django
djangorestframework
psycopg2-binary
drf-yasg
Faker
python-decouple
